Reorganized the core script manager
