+++ miezepy/core/module_process.py
@@ -74,149 +74,7 @@ class Process_MIEZE(Process_Handler):
         '''
         Process_Handler.__init__(self, env)
         self.env = env
-        self.initialize()
-
-    def initialize(self):
-        '''
-        Initialize the default python scipts so that 
-        the system can be set.
-        '''
-        self.default_scripts = []
-        with open(
-            os.path.dirname(os.path.realpath(__file__))
-            + '/process_modules/import_process.py','r') as f:
-            self.default_scripts.append(f.read())
-        with open(
-            os.path.dirname(os.path.realpath(__file__))
-            + '/process_modules/set_fit_process.py','r') as f:
-            self.default_scripts.append(f.read())
-        with open(
-            os.path.dirname(os.path.realpath(__file__))
-            + '/process_modules/phase_process.py','r') as f:
-            self.default_scripts.append(f.read())
-        with open(
-            os.path.dirname(os.path.realpath(__file__))
-            + '/process_modules/reduction_process.py','r') as f:
-            self.default_scripts.append(f.read())
-        with open(os.path.dirname(
-            os.path.realpath(__file__))
-            + '/process_modules/post_process.py','r') as f:
-            self.default_scripts.append(f.read())
-
-        self.editable_scripts = list(self.default_scripts)
-
-    def loadScripts(self, path):
-        '''
-        Initialize the default python scipts so that 
-        the system can be set.
-
-        Parameters
-        ----------
-        path : string
-            The path of the file
-        '''
-        with open(path,'r') as f:
-            text = f.read()
-
-        if 'metadata_class.add_metadata' in text:
-            text = self.reformatScript(text)
-
-        check = text.split('##--FIT-PARA--##')
-        if len(check) > 1:
-            file_type = 'new'
-        else:
-            file_type = 'old'
-
-        if file_type == 'old':
-            strings = [
-                text.split('##--IMPORT--##')[1],
-                text.split('##--PHASE--##')[1].split(
-                    'value = foils_in_echo)')[0]+'value = foils_in_echo)',
-                'environnement = self.env\n'+text.split('##--PHASE--##')[1].split(
-                    'value = foils_in_echo)')[1],
-                text.split('##--REDUCTION--##')[1],
-                text.split('##--POST--##')[1] ]
-        elif file_type == 'new':
-            strings = [
-                text.split('##--IMPORT--##')[1],
-                text.split('##--FIT-PARA--##')[1],
-                text.split('##--PHASE--##')[1],
-                text.split('##--REDUCTION--##')[1],
-                text.split('##--POST--##')[1]]
-
-        self.editable_scripts = list(strings)
-
-    def reformatScript(self, text):
-        '''
-        This function will reformat the text input to meet the
-        new specifications of our software. Most function calls
-        have been updated.
-
-        Parameters
-        ----------
-        text : string
-            The formated text to be formater
-
-        Returns
-        ------- 
-        text : string
-            The formated text to returned to the text importer
-        '''
-        text = text.replace('add_metadata', 'addMetadata')
-        text = text.replace('get_result', 'results.getLastResult')
-        text = text.replace('calculate_echo', 'calculateEcho')
-        text = text.replace('calculate_shift', 'calcShift')
-        text = text.replace('calculate_ref_contrast', 'calcContrastRef')
-        text = text.replace('calculate_contrast', 'calcContrastMain')
-        text = text.replace('environnement.process.remove_foils()\n', '')
-        text = text.replace(
-            "environnement.fit.set_parameter( name = 'foils_in_echo', value = foils_in_echo)\n", 
-            "environnement.fit.set_parameter( name = 'foils_in_echo', value = foils_in_echo)\n"
-            +"environnement.fit.set_parameter( name = 'exposure', value = exposure)\n"
-            +"environnement.instrument.setDetector(instrument, detector)\n")
-        text = text.replace(
-            "environnement.fit.set_parameter( name = 'Select',",
-            "instrument = 'Reseda'\n"
-            +"detector = None\n\n"
-            +"exposure = False\n\n"
-            +"environnement.fit.set_parameter( name = 'Select',")
-
-        return text
-
-    def saveScripts(self, path, strings):
-        '''
-        Initialize the default python scipts so that 
-        the system can be set.
-
-        Parameters
-        ----------
-        path : string
-            The file path to be saved
-
-        strings : string array
-            The list of scripts to be saved
-        '''
-        string = (
-            "##--IMPORT--##\n"
-            + strings[0]
-            + "\n##--IMPORT--##\n"
-            "##--FIT-PARA--##\n"
-            + strings[1]
-            + "\n##--FIT-PARA--##\n"
-            + "##--PHASE--##\n"
-            + strings[2]
-            + "\n##--PHASE--##\n"
-            + "##--REDUCTION--##\n"
-            + strings[3]
-            + "\n##--REDUCTION--##\n"
-            + "##--POST--##\n"
-            + strings[4]
-            + "\n##--POST--##\n")
-
-        f = open(path, 'w')
-        f.write(string)
-        f.close()
-
+        
     def calculateEcho(self):
         '''
         This function will calculate the echo times of
